+++ app/domain/operations/fdsh/vlp/h92/request_initial_verification.rb
@@ -14,43 +14,28 @@ module Operations
           include Dry::Monads[:result, :do, :try]
           include Acapi::Notifiers
 
-          # @param [ Hash ] params Applicant Attributes
-          # @return [ BenefitMarkets::Entities::Applicant ] applicant Applicant
-          def call(params)
-            payload_param = yield construct_payload_hash(params)
-            payload_value = yield validate_payload(payload_param)
-            payload_entity = yield create_payload_entity(payload_value)
-            payload = yield publish(payload_entity)
+          def call(person)
+            payload_entity = yield validate_and_build_payload_entity(person)
+            event  = yield build_event(payload_entity.to_h)
+            result = yield publish(event)
 
-            Success(payload)
+            Success(result)
           end
 
           private
 
-          def construct_payload_hash(person)
-            if person.is_a?(::Person)
-              Operations::Transformers::PersonTo::Cv3Person.new.call(person)
-            else
-              Failure("Invalid Person Object #{person}")
-            end
+          def validate_and_build_payload_entity(person)
+            Operations::Fdsh::ValidatePersonPayload.new.call(person, :vlp)
           end
 
-          def validate_payload(value)
-            result = AcaEntities::Contracts::People::PersonContract.new.call(value)
-            if result.success?
-              Success(result)
-            else
-              hbx_id = value[:hbx_id]
-              Failure("Person with hbx_id #{hbx_id} is not valid due to #{result.errors.to_h}.")
-            end
+          def build_event(payload)
+            event('events.fdsh.vlp.h92.initial_verification_requested', attributes: payload, headers: { correlation_id: payload[:hbx_id] })
           end
 
-          def create_payload_entity(value)
-            Success(AcaEntities::People::Person.new(value.to_h))
-          end
+          def publish(event)
+            event.publish
 
-          def publish(payload)
-            Operations::Fdsh::Vlp::H92::PublishInitialVerificationRequest.new.call(payload)
+            Success("Successfully published the payload to fdsh_gateway")
           end
         end
       end
